+++ TodoList_SwiftUI/Enums/SegmentIndex.swift
@@ -1,0 +1,15 @@
+//
+//  SegmentIndex.swift
+//  TodoList_SwiftUI
+//
+//  Created by 土橋正晴 on 2024/07/21.
+//
+
+import Foundation
+
+enum SegmentIndex: Int, CaseIterable {
+    case all = 0
+    case active = 1
+    case expired = 2
+    case complete = 3
+}
